client/requirements.txt
grpcio~=1.59.3
grpcio-tools~=1.59.3
Pillow~=10.1.0
numpy~=1.26.2
tensorflow~=2.15.0
